Start refactoring of cassandra modules.
Moved Session instantiation code into a separate module, small cleanup
of ApdbCassandra mess.
